+++ src/model/attention/split_cross_attention.py
@@ -89,7 +89,7 @@ class CrossAttention(tf.keras.layers.Layer):
 
         a_cross = tf.nn.softmax(
             tf.matmul(query, tf.transpose(key, perm=[0, 1, 3, 2]))
-            / tf.sqrt(tf.cast(self.embedding_dim, dtype=tf.float32)),
+            / tf.sqrt(tf.cast(self.embedding_dim, dtype=tf.float16)),
             axis=-1,
         )
         o = tf.matmul(a_cross, value)  # (8, 512)
