+++ src/model/loss_functions/class_loss_functions.py
@@ -3,4 +3,7 @@ import tensorflow as tf
 
 @tf.function
 def cls_loss(tgt_labels, pred_cls):
-    return tf.keras.losses.CategoricalCrossentropy()(tgt_labels, pred_cls)
+    return tf.cast(
+        tf.keras.losses.CategoricalCrossentropy()(tgt_labels, pred_cls),
+        dtype=tf.float16,
+    )
