+++ src/train_validate/validate.py
@@ -36,14 +36,14 @@ def validate(model, images, targets, num_cls: int = 8):
     matched_bbox = tf.gather(mini_det_output[..., num_cls:], matched_idx, batch_dims=1)
     matched_bbox = tf.reshape(matched_bbox, shape=(-1, 4))
 
-    mini_det_loss = tf.constant(0.5, dtype=tf.float32) * cls_loss(
+    mini_det_loss = tf.constant(0.5, dtype=tf.float16) * cls_loss(
         tgt_oh_labels, matched_cls
-    ) + tf.constant(0.5, dtype=tf.float32) * boxes_loss_v2(
+    ) + tf.constant(0.5, dtype=tf.float16) * boxes_loss_v2(
         tgt_boxes, matched_bbox, alpha=0.0
     )
-    model_loss = tf.constant(0.5, dtype=tf.float32) * cls_loss(
+    model_loss = tf.constant(0.5, dtype=tf.float16) * cls_loss(
         tgt_oh_labels, pred_cls_flat
-    ) + tf.constant(0.7, dtype=tf.float32) * boxes_loss_v2(
+    ) + tf.constant(0.7, dtype=tf.float16) * boxes_loss_v2(
         tgt_boxes, pred_boxes_flat, alpha=0.0
     )
     box_loss = boxes_loss_v2(tgt_boxes, pred_boxes_flat, alpha=0.0)
